requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.5
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.11.1
annotated-types==0.5.0
APScheduler==3.10.3
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.7.0
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
lxml==4.9.3
magic-filter==1.0.12
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.25.1
opencv-python==4.8.0.74
packaging==23.1
pandas==2.1.1
pathspec==0.11.1
platformdirs==3.9.1
pluggy==1.3.0
pydantic==2.3.0
pydantic_core==2.6.3
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.19
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
yarl==1.9.2
